Change read/command to command/2/3
Read is deprecated in favor of command/2/3 that now always returns a
return value per actor. In this way, the old read and command can be
implemented on top of the new command implementation.

Note: the error `actor_not_active` is no longer possible to get, instead
you will only get an empty list back from command/2/3 if there are no
actors.